requirements_dev.txt
bumpversion==0.5.3
wheel==0.29.0
watchdog==0.8.3
flake8==2.5.4
tox==2.3.1
coverage==4.1
Sphinx==1.4.3

pytest==2.9.2
github3.py==1.0.0a4
